+++ examples/anchors_example.rs
@@ -1,4 +1,7 @@
-use bevy::prelude::*;
+use bevy::{
+    color::palettes::css::{BLUE, GREEN, RED, YELLOW},
+    prelude::*,
+};
 use bevy_inspector_egui::quick::{ResourceInspectorPlugin, WorldInspectorPlugin};
 use bevy_silk::prelude::*;
 
@@ -30,10 +33,10 @@ fn setup(
     });
     let mesh_handle = meshes.add(Cuboid::default());
     [
-        (Color::BLUE, [-10.0, 0.0]),
-        (Color::GREEN, [10.0, 0.0]),
-        (Color::YELLOW, [0.0, -10.0]),
-        (Color::RED, [0.0, 10.0]),
+        (Color::from(BLUE), [-10.0, 0.0]),
+        (Color::from(GREEN), [10.0, 0.0]),
+        (Color::from(YELLOW), [0.0, -10.0]),
+        (Color::from(RED), [0.0, 10.0]),
     ]
     .map(|(color, [x, z])| {
         commands.spawn(PbrBundle {
@@ -59,7 +62,7 @@ fn spawn_cloth(
         .spawn((
             PbrBundle {
                 mesh: anchor_mesh.clone(),
-                material: materials.add(Color::RED),
+                material: materials.add(Color::from(RED)),
                 transform: Transform::from_xyz(-15.0, 15.0, 15.0),
                 ..Default::default()
             },
@@ -70,7 +73,7 @@ fn spawn_cloth(
         .spawn((
             PbrBundle {
                 mesh: anchor_mesh,
-                material: materials.add(Color::GREEN),
+                material: materials.add(Color::from(GREEN)),
                 transform: Transform::from_xyz(15.0, 15.0, 15.0),
                 ..Default::default()
             },