+++ examples/balloon_example.rs
@@ -1,4 +1,7 @@
-use bevy::prelude::*;
+use bevy::{
+    color::palettes::css::{BLUE, GREEN, RED, YELLOW},
+    prelude::*,
+};
 use bevy_inspector_egui::quick::{ResourceInspectorPlugin, WorldInspectorPlugin};
 use bevy_silk::prelude::*;
 
@@ -31,10 +34,10 @@ fn setup(
     commands.spawn(DirectionalLightBundle::default());
     let mesh_handle = meshes.add(Cuboid::default());
     [
-        (Color::BLUE, [-10.0, 0.0]),
-        (Color::GREEN, [10.0, 0.0]),
-        (Color::YELLOW, [0.0, -10.0]),
-        (Color::RED, [0.0, 10.0]),
+        (Color::from(BLUE), [-10.0, 0.0]),
+        (Color::from(GREEN), [10.0, 0.0]),
+        (Color::from(YELLOW), [0.0, -10.0]),
+        (Color::from(RED), [0.0, 10.0]),
     ]
     .map(|(color, [x, z])| {
         commands.spawn(PbrBundle {
@@ -58,7 +61,7 @@ fn spawn_cloth(
     commands.spawn((
         PbrBundle {
             mesh: meshes.add(Sphere::new(5.).mesh().ico(10).unwrap()),
-            material: materials.add(Color::YELLOW),
+            material: materials.add(Color::from(YELLOW)),
             transform: Transform::from_xyz(0.0, 2.0, 0.0),
             ..Default::default()
         },