+++ src/components/mod.rs
@@ -5,5 +5,5 @@ pub mod cloth_builder;
 /// cloth rendering module
 pub mod cloth_rendering;
 /// collider module
-#[cfg(any(feature = "rapier_collisions", feature = "xpbd_collisions"))]
+#[cfg(any(feature = "rapier_collisions", feature = "avian_collisions"))]
 pub mod collider;