+++ src/lib.rs
@@ -6,7 +6,7 @@
 //!
 //! Add `bevy_silk` as a dependency in the `Cargo.toml`
 //!
-//! `bevy_silk = "0.8"`
+//! `bevy_silk = "0.9"`
 //!
 //! Or follow the main git branch
 //!
@@ -24,6 +24,7 @@
 //! | 0.6.0  | 0.11 |
 //! | 0.7.0  | 0.12 |
 //! | 0.8.0  | 0.13 |
+//! | 0.9.0  | 0.14 |
 //!
 //! ### Plugin
 //!