+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 * Physics improvements (#25)
 * Bevy `0.14`
-* Migrated from `bevy_xpbd3d` to `avian3d`
+* Migrated from `bevy_xpbd` to `avian`
+* (**BREAKING**) Renamed `xpbd_collisions` feature to `avian_collisions`
+* Renamed `xpbd_collision_example` to `avian_collision_example`
 
 ## 0.8.0